feat: add server-side rendering of the page /tracks
